fix: fixes and moved imports to server.js outside src
